[Refactor]#107-AddPassageVC 분리
